+++ pyanalyze/arg_spec.py
@@ -83,6 +83,7 @@ from .value import (
     KVPair,
     NewTypeValue,
     SubclassValue,
+    TypedDictEntry,
     TypedDictValue,
     TypedValue,
     TypeVarValue,
@@ -218,6 +219,7 @@ class ClassesSafeToInstantiate(PyObjectSequenceOption[type]):
         Value,
         Extension,
         KVPair,
+        TypedDictEntry,
         asynq.ConstFuture,
         range,
         tuple,
@@ -732,10 +734,10 @@ class ArgSpecCache:
                     SigParameter(
                         key,
                         ParameterKind.KEYWORD_ONLY,
-                        default=None if required else KnownValue(...),
-                        annotation=value,
+                        default=None if entry.required else KnownValue(...),
+                        annotation=entry.typ,
                     )
-                    for key, (required, value) in td_type.items.items()
+                    for key, entry in td_type.items.items()
                 ]
                 if td_type.extra_keys is not None:
                     annotation = GenericValue(