+++ pyanalyze/test_typeddict.py
@@ -2,7 +2,7 @@
 from .implementation import assert_is_value
 from .test_name_check_visitor import TestNameCheckVisitorBase
 from .test_node_visitor import assert_passes
-from .value import AnySource, AnyValue, TypedDictValue, TypedValue
+from .value import AnySource, AnyValue, TypedDictEntry, TypedDictValue, TypedValue
 
 
 class TestExtraKeys(TestNameCheckVisitorBase):
@@ -21,7 +21,7 @@ class TestExtraKeys(TestNameCheckVisitorBase):
             assert_is_value(
                 x,
                 TypedDictValue(
-                    {"a": (True, TypedValue(str))}, extra_keys=TypedValue(int)
+                    {"a": TypedDictEntry(TypedValue(str))}, extra_keys=TypedValue(int)
                 ),
             )
 
@@ -79,7 +79,7 @@ class TestExtraKeys(TestNameCheckVisitorBase):
     def test_compatibility(self):
         from typing import Any, Dict
 
-        from typing_extensions import TypedDict
+        from typing_extensions import ReadOnly, TypedDict
 
         from pyanalyze.extensions import has_extra_keys
 
@@ -95,14 +95,26 @@ class TestExtraKeys(TestNameCheckVisitorBase):
         class TD3(TypedDict):
             a: str
 
+        @has_extra_keys(ReadOnly[int])
+        class TD4(TypedDict):
+            a: str
+
         def want_td(td: TD) -> None:
+            pass
+
+        def want_td4(td: TD4) -> None:
             pass
 
         def capybara(td: TD, td2: TD2, td3: TD3, anydict: Dict[str, Any]) -> None:
             want_td(td)
-            want_td(td2)
+            want_td(td2)  # E: incompatible_argument
             want_td(td3)  # E: incompatible_argument
             want_td(anydict)
+
+            want_td4(td)
+            want_td4(td2)
+            want_td4(td3)  # E: incompatible_argument
+            want_td4(anydict)
 
     @assert_passes()
     def test_iteration(self):
@@ -130,7 +142,7 @@ class TestExtraKeys(TestNameCheckVisitorBase):
                 assert_type(k, str)
                 assert_type(v, str)
             for k in td2:
-                assert_type(k, Literal["a"])
+                assert_type(k, Union[str, Literal["a"]])
 
 
 class TestTypedDict(TestNameCheckVisitorBase):
@@ -151,7 +163,10 @@ class TestTypedDict(TestNameCheckVisitorBase):
             assert_is_value(
                 cap,
                 TypedDictValue(
-                    {"x": (True, TypedValue(int)), "y": (True, TypedValue(str))}
+                    {
+                        "x": TypedDictEntry(TypedValue(int)),
+                        "y": TypedDictEntry(TypedValue(str)),
+                    }
                 ),
             )
             Capybara(x=1)  # E: incompatible_call
@@ -160,7 +175,10 @@ class TestTypedDict(TestNameCheckVisitorBase):
             assert_is_value(
                 maybe_cap,
                 TypedDictValue(
-                    {"x": (True, TypedValue(int)), "y": (False, TypedValue(str))}
+                    {
+                        "x": TypedDictEntry(TypedValue(int)),
+                        "y": TypedDictEntry(TypedValue(str), required=False),
+                    }
                 ),
             )
 
@@ -187,14 +205,20 @@ class TestTypedDict(TestNameCheckVisitorBase):
             assert_is_value(
                 x,
                 TypedDictValue(
-                    {"a": (True, TypedValue(int)), "b": (True, TypedValue(str))}
+                    {
+                        "a": TypedDictEntry(TypedValue(int)),
+                        "b": TypedDictEntry(TypedValue(str)),
+                    }
                 ),
             )
             assert_is_value(x["a"], TypedValue(int))
             assert_is_value(
                 y,
                 TypedDictValue(
-                    {"a": (True, TypedValue(int)), "b": (True, TypedValue(str))}
+                    {
+                        "a": TypedDictEntry(TypedValue(int)),
+                        "b": TypedDictEntry(TypedValue(str)),
+                    }
                 ),
             )
             assert_is_value(y["a"], TypedValue(int))
@@ -242,3 +266,108 @@ class TestTypedDict(TestNameCheckVisitorBase):
             f({"a": "a"})  # E: incompatible_argument
             g({"c": 1.0})
             g({})  # E: incompatible_argument
+
+
+class TestReadOnly(TestNameCheckVisitorBase):
+    @assert_passes()
+    def test_basic(self):
+        from typing import Any, Dict
+
+        from typing_extensions import NotRequired, ReadOnly, TypedDict
+
+        class TD(TypedDict):
+            a: ReadOnly[NotRequired[int]]
+            b: ReadOnly[str]
+
+        def capybara(td: TD, anydict: Dict[str, Any]) -> None:
+            td["a"] = 1  # E: readonly_typeddict
+            td["b"] = "a"  # E: readonly_typeddict
+            td.update(anydict)  # E: invalid_typeddict_key
+            td.setdefault("a", 1)  # E: readonly_typeddict
+            td.setdefault("b", "a")  # E: readonly_typeddict
+            td.pop("a")  # E: readonly_typeddict
+            td.pop("b")  # E: incompatible_argument
+            del td["a"]  # E: readonly_typeddict
+            del td["b"]  # E: incompatible_argument
+
+    @assert_passes()
+    def test_compatibility(self):
+        from typing import Any, Dict
+
+        from typing_extensions import ReadOnly, TypedDict
+
+        class TD(TypedDict):
+            a: int
+
+        class TD2(TypedDict):
+            a: ReadOnly[int]
+
+        class TD3(TypedDict):
+            a: bool
+
+        class TD4(TypedDict):
+            a: str
+
+        def want_td(td: TD) -> None:
+            pass
+
+        def want_td2(td: TD2) -> None:
+            pass
+
+        def capybara(
+            td: TD, td2: TD2, td3: TD3, td4: TD4, anydict: Dict[str, Any]
+        ) -> None:
+            want_td(td)
+            want_td(td2)  # E: incompatible_argument
+            want_td(td3)  # E: incompatible_argument
+            want_td(td4)  # E: incompatible_argument
+            want_td(anydict)
+
+            want_td2(td)
+            want_td2(td2)
+            want_td2(td3)
+            want_td2(td4)  # E: incompatible_argument
+            want_td2(anydict)
+
+
+class TestClosed(TestNameCheckVisitorBase):
+    @assert_passes()
+    def test_basic(self):
+        from typing import Any, Dict
+
+        from typing_extensions import NotRequired, TypedDict
+
+        class Closed(TypedDict, closed=True):
+            a: NotRequired[int]
+            b: str
+
+        class Open(TypedDict):
+            a: NotRequired[int]
+            b: str
+
+        def want_closed(td: Closed) -> None:
+            pass
+
+        def want_open(td: Open) -> None:
+            pass
+
+        def capybara(closed: Closed, open: Open, anydict: Dict[str, Any]) -> None:
+            closed["a"] = 1
+            closed["b"] = "a"
+            closed["a"] = "x"  # E: incompatible_argument
+
+            open["a"] = 1
+            open["b"] = "a"
+            open["a"] = "x"  # E: incompatible_argument
+
+            closed.update(anydict)  # E: invalid_typeddict_key
+            open.update(anydict)  # E: invalid_typeddict_key
+
+            x: Closed = {"a": 1, "b": "a", "c": "x"}  # E: incompatible_assignment
+            y: Open = {"a": 1, "b": "a", "c": "x"}
+
+            want_closed(closed)
+            want_closed(open)  # E: incompatible_argument
+
+            want_open(open)
+            want_open(closed)